Separate tests by endpoint
